fix: build 에러 해결
