+++ src/app/providers.tsx
@@ -1,11 +1,23 @@
-// app/providers.tsx
+//app/provider.tsx
 "use client";
 
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
-import { ReactNode, useState } from "react";
+import { useState } from "react";
 
-export default function Providers({ children }: { children: ReactNode }) {
-  const [client] = useState(() => new QueryClient());
+export default function Providers({ children }: { children: React.ReactNode }) {
+  const [queryClient] = useState(
+    () =>
+      new QueryClient({
+        defaultOptions: {
+          queries: {
+            staleTime: 60 * 1000, // 1분
+            retry: 1,
+          },
+        },
+      })
+  );
 
-  return <QueryClientProvider client={client}>{children}</QueryClientProvider>;
+  return (
+    <QueryClientProvider client={queryClient}>{children}</QueryClientProvider>
+  );
 }
