--- src/app/providers/ReactQueryProvider.tsx
@@ -1,12 +1,0 @@
-"use client";
-
-import { QueryClient, QueryClientProvider } from "react-query";
-import { ReactNode, useState } from "react";
-
-export function ReactQueryProvider({ children }: { children: ReactNode }) {
-  const [queryClient] = useState(() => new QueryClient());
-
-  return (
-    <QueryClientProvider client={queryClient}>{children}</QueryClientProvider>
-  );
-}
